Solution for exercise 4.21*: bloglist expansion, step9. Added authentication for deleting.

+++ controllers/blogs.js
@@ -47,8 +47,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+  const currentUserId = decodedToken.id
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (!(currentUserId === blogToDelete.user.toJSON())) {
+    response.status(401).json({ error: 'invalid user' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
